finalise authentification on all pages
